+++ tests/fixtures/noqa/noqa.py
@@ -596,6 +596,10 @@ a_list[1:3] = [1, 2]  # noqa: WPS362
 a_list[slice(1)] = [1, 2]  # noqa: WPS362
 
 
+def function_with_systemerror():
+    raise SystemError(1)  # noqa: WPS363
+
+
 try:
     cause_errors()
 except ValueError or TypeError:  # noqa: WPS455